backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.4
langchain==0.1.6
langchain-groq==0.0.1
langgraph==0.0.20
groq==0.4.2
sentence-transformers==2.3.1
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.3
hypothesis==6.92.1
